+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
 
+## v1.1.0
+change: findPaginated was changed to findInfinite
+fix: fixing docs on comments
+new: New APIs findNext and findPrevious
+
 ## v1.0.0
 Initial release